Modified MySQL database, entities and repositories
