--- src/main/java/com/revature/annotations/AuthRestriction.java
@@ -1,5 +1,0 @@
-package com.revature.annotations;
-
-public enum AuthRestriction {
-    LoggedIn
-}
