--- src/main/java/com/revature/annotations/Authorized.java
@@ -1,12 +1,0 @@
-package com.revature.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Authorized {
-    public AuthRestriction value() default AuthRestriction.LoggedIn;
-}
