--- src/main/java/com/revature/dtos/LoginRequest.java
@@ -1,14 +1,0 @@
-package com.revature.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginRequest {
-
-    private String email;
-    private String password;
-}
