--- src/main/java/com/revature/dtos/ProductInfo.java
@@ -1,21 +1,0 @@
-package com.revature.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductInfo {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private int quantity;
-}
