+++ src/main/java/com/revature/ecommerce/SpringBootEcommerceApplication.java
@@ -1,0 +1,13 @@
+package com.revature.ecommerce;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootEcommerceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBootEcommerceApplication.class, args);
+	}
+
+}
