+++ src/main/java/com/revature/ecommerce/annotations/AuthRestriction.java
@@ -1,0 +1,5 @@
+package com.revature.ecommerce.annotations;
+
+public enum AuthRestriction {
+    LoggedIn
+}
