+++ src/main/java/com/revature/ecommerce/annotations/Authorized.java
@@ -1,0 +1,12 @@
+package com.revature.ecommerce.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Authorized {
+    public AuthRestriction value() default AuthRestriction.LoggedIn;
+}
