+++ src/main/java/com/revature/ecommerce/config/MyDataRestConfig.java
@@ -1,0 +1,69 @@
+package com.revature.ecommerce.config;
+
+import com.revature.ecommerce.models.Product;
+import com.revature.ecommerce.models.ProductCategory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
+import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurer;
+import org.springframework.http.HttpMethod;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+
+import javax.persistence.EntityManager;
+import javax.persistence.metamodel.EntityType;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+
+@Configuration
+public class MyDataRestConfig implements RepositoryRestConfigurer {
+
+    private EntityManager entityManager;
+
+    @Autowired
+    public MyDataRestConfig(EntityManager theEntityManager){
+        entityManager = theEntityManager;
+    }
+
+    @Override
+    public void configureRepositoryRestConfiguration(RepositoryRestConfiguration config, CorsRegistry cors) {
+
+//        HttpMethod[] theUnsupportedActions = {HttpMethod.PUT, HttpMethod.POST, HttpMethod.DELETE};
+//
+//        //disable Http methods for Products: put, post, and delete
+//        config.getExposureConfiguration()
+//                .forDomainType(Product.class)
+//                .withItemExposure((metdata, httpMethods) -> httpMethods.disable(theUnsupportedActions))
+//                .withCollectionExposure((metdata, httpMethods) -> httpMethods.disable(theUnsupportedActions));
+//
+//        //disable Http methods for ProductCategory: put, post, and delete
+//        config.getExposureConfiguration()
+//                .forDomainType(ProductCategory.class)
+//                .withItemExposure((metdata, httpMethods) -> httpMethods.disable(theUnsupportedActions))
+//                .withCollectionExposure((metdata, httpMethods) -> httpMethods.disable(theUnsupportedActions));
+
+        //call an internal helper method
+        exposeIds(config);
+    }
+
+    private void exposeIds(RepositoryRestConfiguration config) {
+        //expose entity ids
+
+        //get list of all entity classes from entity manager
+        Set<EntityType<?>> entities = entityManager.getMetamodel().getEntities();
+
+        //create an array of the entity types
+        List<Class> entityClasses = new ArrayList<>();
+
+        //get entity types for the entities
+        for (EntityType tempEntityType : entities) {
+            entityClasses.add(tempEntityType.getJavaType());
+        }
+
+        //expose the entity ids for the array of entity/domain types
+        Class[] domainTypes = entityClasses.toArray(new Class[0]);
+        config.exposeIdsFor(domainTypes);
+    }
+
+
+}
