+++ src/main/java/com/revature/ecommerce/controllers/AuthController.java
@@ -1,0 +1,57 @@
+package com.revature.ecommerce.controllers;
+
+import com.revature.ecommerce.dtos.LoginRequest;
+import com.revature.ecommerce.dtos.RegisterRequest;
+import com.revature.ecommerce.dtos.LoginRequest;
+import com.revature.ecommerce.dtos.RegisterRequest;
+import com.revature.ecommerce.models.User;
+import com.revature.ecommerce.services.AuthService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpSession;
+import java.util.Optional;
+
+@RestController
+@RequestMapping("/auth")
+@CrossOrigin(origins = {"http://localhost:4200", "http://localhost:3000"}, allowCredentials = "true")
+public class AuthController {
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<User> login(@RequestBody LoginRequest loginRequest, HttpSession session) {
+        Optional<User> optional = authService.findByCredentials(loginRequest.getEmail(), loginRequest.getPassword());
+
+        if(!optional.isPresent()) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        session.setAttribute("user", optional.get());
+
+        return ResponseEntity.ok(optional.get());
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(HttpSession session) {
+        session.removeAttribute("user");
+
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/register")
+    public ResponseEntity<User> register(@RequestBody RegisterRequest registerRequest) {
+        User created = new User(0,
+                registerRequest.getEmail(),
+                registerRequest.getPassword(),
+                registerRequest.getFirstName(),
+                registerRequest.getLastName());
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(authService.register(created));
+    }
+}
