+++ src/main/java/com/revature/ecommerce/dtos/LoginRequest.java
@@ -1,0 +1,14 @@
+package com.revature.ecommerce.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginRequest {
+
+    private String email;
+    private String password;
+}
