+++ src/main/java/com/revature/ecommerce/dtos/RegisterRequest.java
@@ -1,0 +1,15 @@
+package com.revature.ecommerce.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class RegisterRequest {
+    private String email;
+    private String password;
+    private String firstName;
+    private String lastName;
+}
