+++ src/main/java/com/revature/ecommerce/exceptions/NotLoggedInException.java
@@ -1,0 +1,25 @@
+package com.revature.ecommerce.exceptions;
+
+// Exception to be thrown by the AuthAspect
+// Will be handled by a Spring Exception Handler to return a 401
+public class NotLoggedInException extends RuntimeException {
+
+    public NotLoggedInException() {
+    }
+
+    public NotLoggedInException(String message) {
+        super(message);
+    }
+
+    public NotLoggedInException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotLoggedInException(Throwable cause) {
+        super(cause);
+    }
+
+    public NotLoggedInException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
