+++ src/main/java/com/revature/ecommerce/repositories/ProductCategoryRepository.java
@@ -1,0 +1,11 @@
+package com.revature.ecommerce.repositories;
+
+import com.revature.ecommerce.models.ProductCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+import org.springframework.web.bind.annotation.CrossOrigin;
+
+@CrossOrigin("http://localhost:4200")
+@RepositoryRestResource(collectionResourceRel = "productCategory", path = "product-category")
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
+}
