+++ src/main/java/com/revature/ecommerce/repositories/ProductRepository.java
@@ -1,0 +1,17 @@
+package com.revature.ecommerce.repositories;
+
+import com.revature.ecommerce.models.Product;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@CrossOrigin("http://localhost:4200")
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+    Page<Product> findByCategoryId(@RequestParam("id") Long id, Pageable pageable);
+
+    Page<Product> findByNameContaining(@RequestParam("name") String name, Pageable pageable);
+
+}
