+++ src/main/java/com/revature/ecommerce/services/AuthService.java
@@ -1,0 +1,24 @@
+package com.revature.ecommerce.services;
+
+import com.revature.ecommerce.models.User;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class AuthService {
+
+    private final UserService userService;
+
+    public AuthService(UserService userService) {
+        this.userService = userService;
+    }
+
+    public Optional<User> findByCredentials(String email, String password) {
+        return userService.findByCredentials(email, password);
+    }
+
+    public User register(User user) {
+        return userService.save(user);
+    }
+}
