+++ src/main/java/com/revature/ecommerce/services/ProductService.java
@@ -1,0 +1,40 @@
+package com.revature.ecommerce.services;
+
+import com.revature.ecommerce.dtos.ProductInfo;
+import com.revature.ecommerce.models.Product;
+import com.revature.ecommerce.repositories.ProductRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class ProductService {
+
+    private final ProductRepository productRepository;
+
+    public ProductService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    public List<Product> findAll() {
+        return productRepository.findAll();
+    }
+
+    //this int to long
+    public Optional<Product> findById(long id) {
+        return productRepository.findById(id);
+    }
+
+    public Product save(Product product) {
+        return productRepository.save(product);
+    }
+    
+    public List<Product> saveAll(List<Product> productList, List<ProductInfo> metadata) {
+    	return productRepository.saveAll(productList);
+    }
+    //this int to long
+    public void delete(long id) {
+        productRepository.deleteById(id);
+    }
+}
