--- src/main/java/com/revature/models/Product.java
@@ -1,26 +1,0 @@
-package com.revature.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Data
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private int quantity;
-    private double price;
-    private String description;
-    private String image;
-    private String name;
-}
