--- src/main/java/com/revature/models/User.java
@@ -1,23 +1,0 @@
-package com.revature.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Data
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private String email;
-    private String password;
-    private String firstName;
-    private String lastName;
-}
