--- src/main/java/com/revature/repositories/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.revature.repositories;
-
-import com.revature.models.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-}
