--- src/main/java/com/revature/repositories/UserRepository.java
@@ -1,11 +1,0 @@
-package com.revature.repositories;
-
-import com.revature.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    Optional<User> findByEmailAndPassword(String email, String password);
-}
