--- src/test/java/com/revature/ECommerceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.revature;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ECommerceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
