+++ src/test/java/com/revature/ecommerce/SpringBootEcommerceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.revature.ecommerce;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBootEcommerceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
